backend/requirements.txt
django
djangorestframework
black
flake8
isort
